Fix all phpcs and phpstan violations; add pre-commit lint gate
- Exclude phpstan-bootstrap.php from phpcs scanning in all three plugins
- Add phpstan-bootstrap.php stubs for plugin constants and optional
  vs-secrets-manager function/class stubs in vs-mailer and social-feed
- Set phpstan memory-limit=1G and maximumNumberOfProcesses=1 to avoid
  child process OOM crashes; ignoreErrors for AWS SDK optional dependency
- Add missing phpcs rule exclusions (direct DB, file ops, wp_redirect,
  SafeRedirect, unused hook params, timestamp, CommentedOutCode) to
  social-feed/phpcs.xml; mirror phpstan-bootstrap exclude to all three
- Fix phpcs violations: alignment (phpcbf), urlencode→rawurlencode,
  0==float→0.0===float, NonceVerification phpcs:ignore on display-only
  GET reads in admin views
- Fix phpstan violation: VS_Secrets_Manager_Secret_Manager::delete()
  missing from social-feed bootstrap stub
- Add .git/hooks/pre-commit that runs composer phpcs + phpstan on all
  three plugins and blocks the commit if any check fails

+++ social-feed/includes/class-shortcode.php
@@ -237,15 +237,15 @@ class Social_Feed_Shortcode {
 		}
 		if ( $diff < 3600 ) {
 			/* translators: %d: number of minutes */
-			return sprintf( _n( '%d minute ago', '%d minutes ago', floor( $diff / 60 ), 'social-feed' ), floor( $diff / 60 ) );
+			return sprintf( _n( '%d minute ago', '%d minutes ago', (int) floor( $diff / 60 ), 'social-feed' ), (int) floor( $diff / 60 ) );
 		}
 		if ( $diff < 86400 ) {
 			/* translators: %d: number of hours */
-			return sprintf( _n( '%d hour ago', '%d hours ago', floor( $diff / 3600 ), 'social-feed' ), floor( $diff / 3600 ) );
+			return sprintf( _n( '%d hour ago', '%d hours ago', (int) floor( $diff / 3600 ), 'social-feed' ), (int) floor( $diff / 3600 ) );
 		}
 		if ( $diff < 604800 ) {
 			/* translators: %d: number of days */
-			return sprintf( _n( '%d day ago', '%d days ago', floor( $diff / 86400 ), 'social-feed' ), floor( $diff / 86400 ) );
+			return sprintf( _n( '%d day ago', '%d days ago', (int) floor( $diff / 86400 ), 'social-feed' ), (int) floor( $diff / 86400 ) );
 		}
 
 		return date_i18n( get_option( 'date_format' ), $time );
